+++ app/controller/UserController.php
@@ -8,6 +8,10 @@ use App\Services\UserService;
 use App\Collection\UserCollection;
 use App\Model\User;
 
+/**
+ * Classe responsável por interagir entre a view
+ * e os dados/lógica de usuários.
+ */
 class UserController extends Controller
 {
 	protected $services = array();
@@ -34,12 +38,19 @@ class UserController extends Controller
 		];
 	}
 
+	/**
+	 * Filtro pré aplicado antes de redirecionar solicitação
+	 * a um método do controller.
+	 */
 	public function run($route)
 	{
 		if ($this->services->auth->isSignedIn())
 		{
 			return true;
 		}
+
+		// Caso o usuário não esteja autenticado, o redirecionamos
+		// à página de login.
 
 		app()->module('router')->redirect('login');
 
@@ -48,22 +59,51 @@ class UserController extends Controller
 
 	public function index()
 	{
-		$users = $this->collection->user->getAll();
-		app()->module('view')->load('users/index', ['users' => $users]);
+		if ($this->checkUserPermission('ver_usuários'))
+		{
+			$users = $this->collection->user->getAll();
+
+			app()->module('view')->load('users/index', ['users' => $users]);
+		}
 	}
 
+	/**
+	 * Página de adicionar usuário
+	 */
 	public function create()
 	{
-		app()->module('view')->load('users/create');
+		if ($this->checkUserPermission('criar_usuários'))
+		{
+			app()->module('view')->load('users/create');
+		}
 	}
 
+	/**
+	 * Salvar usuário no banco de dados
+	 */
 	public function store()
 	{
-		$email = filter_input(INPUT_POST, 'email');
-		$password = filter_input(INPUT_POST, 'password');
+		if (!$this->checkUserPermission('criar_usuários'))
+		{
+			return;
+		}
+
+		$args = filter_input_array(INPUT_POST, array(
+			'email' => FILTER_DEFAULT,
+			'password' => FILTER_DEFAULT,
+			'permissions' => array(
+				'filter' => FILTER_SANITIZE_STRING,
+				'flags' => FILTER_REQUIRE_ARRAY
+			)
+		));
+
+		$email = $args['email'];
+		$password = $args['password'];
+		$permissions = $args['permissions'];
 
 		$user = new User();
 		$user->setEmail($email);
+		$user->setPermission($permissions);
 
 		if ($this->services->user->create($user, $password))
 		{
@@ -73,12 +113,21 @@ class UserController extends Controller
 		else
 		{
 			app()->module('session')->flash('error', $this->services->user->message);
+			app()->module('session')->flash('input_email', $email);
 			app()->module('router')->redirect('users.create');
 		}
 	}
 
+	/**
+	 * Atualizar usuário no banco de dados
+	 */
 	public function update($id)
 	{
+		if (!$this->checkUserPermission('editar_usuários'))
+		{
+			return;
+		}
+
 		$user = new User();
 		$user->setId($id);
 		$user->setName(filter_input(INPUT_POST, 'name'));
@@ -89,11 +138,48 @@ class UserController extends Controller
 		echo 'atualizar usuário';
 	}
 
+	/**
+	 * Remover usuário do banco de dados
+	 */
 	public function destroy($id)
 	{
+		if (!$this->checkUserPermission('remover_usuários'))
+		{
+			return;
+		}
+
+		// TODO: implementar deletar usuário
+
 		$result = $this->services->user->delete($id);
 
 		echo 'deletar usuário';
+	}
+
+	/**
+	 * Verifica se usuário atual possui permissão.
+	 * 
+	 * @param string $permission 		Permissão para verificar
+	 * @return bool
+	 */
+	protected function checkUserPermission($permission)
+	{
+		$user = $this->services->user->findById($this->services->auth->getCurrentUserId());
+
+		if (in_array($permission, $user->getPermission()))
+		{
+			return true;
+		}
+
+		// Caso o usuário não tenha a permissão, o redirecionamos
+		// à página principal do painel, com mensagem de erro
+
+		http_response_code(403);
+
+		app()->module('session')->flash('error', 'Você não possui permissões suficientes para acessar esta página.');
+
+		app()->module('router')->redirect('index');
+
+		return false;
 	}
 }
